refactor: move CKAN/SDMX detection to plugins, scout re-imports
